update notification interface
