+++ src/components/beta/layout/mobile/notification_item_mobile.tsx
@@ -1,4 +1,4 @@
-import { INotification, NOTIFICATION_TYPE } from '@/constants/notification';
+import { INotification, NOTIFICATION_TYPE } from '@/interfaces/notification';
 import { cn } from '@/lib/utils/common';
 
 interface NotificationItemMobileProps {
